Begun making essential classes
I have begun to make all of the really basic and required classes for the game in order for it to meets its basic requirements. The game is not in a runnable state and will merge to master when it is in such a state.

So far, basic Ship class, basic Board class and basic Square class are really shaping up, as well as a couple other small classes. The main start-up class needs work but that will be continued later.