Got music playing working
Playing music in the application now works such that on application startup music begins playing right away and the user enjoys it.
Also added a song for the user to enjoy.

Will likely add some sort of settings for the music in the future so the user isn't forced to listen to it.