Refactored code and added more hitting logic
Refactored the packages in the code so that the game elements are
seperate from everything else.

Also make a fill animation for when the user hits a ship with cannon
fire so that a red fill is seen briefly, much like the yellow fill they
see when they miss the ship.