Made a main menu button in the game, fixed game bug
There was a bug in the game which counted down your ship's moves even if
the ship didn't actually move, so I fixed this.

Also made a button in the game that allowed the user to return to the
main menu. Considering removing menus entirely because it isn't
necessary for this game.

I also changed the number of moves per turn to four to increase the
amount of movement per turn.