Got battle mode working
Got hte battle mode between ships working. Basically, if two ships move
to the same square, battle mode will be entered, and in battle mode the
players have to both rush to the keyboard and press their corresponding
select key more than the other player to win the battle.

Some changes were made to the overall framework (for the better really)
to make this mode work. Now that this has been implemented, ship
movement is complete.