Completed key masher, began to refactor for .jar
Completed the key masher feature and also realized that the menu spawns
a new game without terminating the old one, if there is one, so now I've
fixed that.

I've also begun to rework the code so that it will work in a .jar file,
something which I'm almost done doing, just need to figure out the
music.