Added help, changed menu layout.
Added a help button to the menu which displays a dialog with help for
the user.

Also changed the layout of the menu so that all the buttons are in a row
with each other at the top of the screen, so there's no sidebar anymore.
Preparing for final submission, still need to delete the commented out
code.